fix(wine-lists): audit fixes — cellar-cascade logo cleanup, shared logo helper, branding null guard
- Cellar deletion hard-deletes WineList docs; unlink their logo files first
  (the docs hold the only references — orphans stayed publicly fetchable).
- Extract LOGO_DIR/deleteLogoFile into services/wineListLogos.js and use it
  from the routes, the cellar cascade and the GDPR purge (was inlined twice).
  Directory creation is now lazy so requiring the module in tests never
  touches the filesystem.
- Guard the PUT logoUrl restore against `branding: null` (TypeError -> 500).
